web/flaskDemo/requirements.txt
# flask-restful base
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3

# casbin
flask-authz==2.5.0
casbin-sqlalchemy-adapter==0.5.0